add shellcheck workflow and fix shellcheck errors
This adds a shellcheck github action and a companion dependabot action to keep
it--and future Github Actions updated.

This also edits one of the first shell scripts I saw, to show some of the errors
it can catch, which will provide an overall safer and more consistent experience
with shell scripts that drive any part of the software.

Signed-off-by: Jacob Salmela <me@jacobsalmela.com>